Cleanup X2CPG API
